Document Retrieval using queries and structs.
The bucketPointer struct serves as a pointer to the bucket containing
the words that are included in that bucket. The LinkNode struct contains
information for the document identification number (in order to handle
multiple documents), a pointer to each individual word in a document, as
well as a pointer to the next node. Finally, the LinkList struct
resembles the creation of a Linked List in C. In this particular case,
the struct contains two pointers, bucketHead and bucketTail, which
represent the beginning and ending of the list, along with the total
size of the list. With these parameters known, every element in the list
can be accessed by adjusting bucketHead and/or bucketTail. The first
function implemented is hash_code(w) which takes in a string as input,
and uses ASCII values that are coded to each letter and converts each
letter into a numerical value. The resulting value is added to the
number of buckets which is specified by the user (to prove that the
algorithm works in all scenarios). The hash_table_insert() function
takes in the bucketPointer struct, the documentation identification, and
a word contained within a document. When a corresponding word is given a
hashing value, it is paired with the identification number, which is
individually incremented during each step of the way. The function
ensures that the document is not added multiple times in any bucket for
an identical string. As long as the recurrence factor is not repeated,
the function will call add_to_list() and the information is passed
along. This function inserts the information (the string and
corresponding ID number) into the hash table, done by allocating memory
for data to be stored in the node. When search_list()  is called, the
list is searched for a word that matches a keyword (or multiple
depending on what the user chooses to do). If a match is found, a result
of 1 signifies a success, whereas a value of 0 refers to a complete
mismatch. The insert_list method adds an element into the linked list as
long as an element is free in space somewhere along the way. The
training() method is perhaps the most essential function in the program
as it handles all the pre-processing. It is built off of the
bucketPointer struct defined earlier. It inputs three text files (or
documents) and obtains the hashing value for each character as well as
its searching value. After all of the files are closed, the entire
hashing table has been returned. The read_query() function obtains input
from the user, asking for words and phrases to search for within the
three documents being read. Memory is allocated in order to perform this
task (the string the user inputs, which is then returned). The score()
function maintains the number of matches found in the documents based
upon the keywords that the user has currently entered. The query word is
matched up against the current node (which is adjusted at each point)
for each position of the list for all documents. The function
find_match() returns the specific documents which consist of the
keyphrases that the user has entered. If there is a direct correlation
between what the user has asked the document search versus what words
are contained within the document, a “perfect match” is recorded. The
overall score of the Keyword to document is printed out to the screen in
print(), with again a “perfect match” notification when conditions
apply. Finally, main() brings all the functions together. Training() is
called to read the documents and develop the hash table, with the number
of buckets specified by the user. The keywords are then searched for and
the locations of these relative words if found in the documents are
tracked. When the user inputs the query set, the string is read, and the
corresponding matching score is calculated, with higher scores ordered
before lower ones (by relevance). The user is prompted to enter another
query set if desired, otherwise entering an # will cause the program to
exit.